services/ml/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "clara-ml"
version = "0.1.0"
description = "CLARA ML/NLP P0 skeleton with LangChain/LangGraph and Vietnamese NLP baseline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "CLARA Team" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "PyYAML>=6.0.2",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "numpy>=2.1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/clara_ml"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
